requirements.txt
absl-py==2.2.2
accelerate==1.7.0
aiohappyeyeballs==2.6.1
aiohttp==3.8.1
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
astunparse==1.6.3
async-timeout==4.0.3
attrs==25.3.0
beautifulsoup4==4.13.4
Brotli==1.0.9
cachetools==5.5.2
cbor==1.0.0
certifi==2025.4.26
charset-normalizer==2.1.1
click==8.1.8
colorama==0.4.6
datasets==2.19.0
dill==0.3.7
distro==1.9.0
exceptiongroup==1.3.0
faiss-cpu==1.11.0
filelock==3.17.0
FlagEmbedding==1.3.2
flatbuffers==25.2.10
frozenlist==1.6.0
fsspec==2024.3.1
gast==0.6.0
gmpy2==2.2.1
google-auth==2.40.1
google-auth-oauthlib==1.2.2
google-pasta==0.2.0
grpcio==1.71.0
h11==0.16.0
h5py==3.13.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.32.0
idna==3.7
ijson==3.4.0
importlib_metadata==8.7.0
inscriptis==2.6.0
ir_datasets==0.5.10
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
keras==2.15.0
libclang==18.1.1
lxml==5.4.0
lz4==4.4.4
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mkl_fft==1.3.11
mkl_random==1.2.8
mkl-service==2.4.0
ml-dtypes==0.2.0
mpmath==1.3.0
multidict==6.4.4
multiprocess==0.70.15
namex==0.0.9
networkx==3.2.1
numpy==1.26.4
oauthlib==3.2.2
openai==1.82.0
opt_einsum==3.4.0
optree==0.15.0
packaging==25.0
pandas==2.2.3
peft==0.15.2
pillow==11.1.0
pip==25.1
propcache==0.3.1
protobuf==4.25.7
psutil==7.0.0
pyarrow==20.0.0
pyarrow-hotfix==0.7
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.5
pydantic_core==2.33.2
Pygments==2.19.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rich==14.0.0
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.13.1
sentence-transformers==4.1.0
sentencepiece==0.2.0
setuptools==78.1.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
sympy==1.13.1
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.1.0
threadpoolctl==3.6.0
tokenizers==0.19.1
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tqdm==4.67.1
transformers==4.44.2
trec-car-tools==2.6
typing_extensions==4.12.2
typing-inspection==0.4.1
tzdata==2025.2
unlzw3==0.2.3
urllib3==2.3.0
warc3-wet==0.2.5
warc3-wet-clueweb09==0.2.5
Werkzeug==3.1.3
wheel==0.45.1
win-inet-pton==1.1.0
wrapt==1.14.1
xxhash==3.5.0
yarl==1.20.0
zipp==3.21.0
zlib-state==0.1.9